style: fix oxlint warnings — split component, reduce params
Extract CompactCrateSection to separate file (fixes no-multi-comp and max-lines-per-function). Combine sharedState booleans into opts object (fixes max-params).
